requirements.txt
requests>=2.28.0
mysql-connector-python>=8.0.0
Pillow>=9.0.0
selenium>=4.0.0
pyautogui>=0.9.53
pytesseract>=0.3.10
opencv-python>=4.5.0
numpy>=1.21.0
python-dotenv>=1.0.0
paramiko>=3.3.0
beautifulsoup4>=4.12.0
